+++ api/server/router/swarm/helpers_test.go
@@ -17,7 +17,8 @@ func TestAdjustForAPIVersion(t *testing.T) {
 	spec := &swarm.ServiceSpec{
 		TaskTemplate: swarm.TaskSpec{
 			ContainerSpec: &swarm.ContainerSpec{
-				Sysctls: expectedSysctls,
+				Sysctls:   expectedSysctls,
+				PidsLimit: 300,
 				Privileges: &swarm.Privileges{
 					CredentialSpec: &swarm.CredentialSpec{
 						Config: "someconfig",
@@ -49,9 +50,16 @@ func TestAdjustForAPIVersion(t *testing.T) {
 	// first, does calling this with a later version correctly NOT strip
 	// fields? do the later version first, so we can reuse this spec in the
 	// next test.
-	adjustForAPIVersion("1.40", spec)
+	adjustForAPIVersion("1.41", spec)
 	if !reflect.DeepEqual(spec.TaskTemplate.ContainerSpec.Sysctls, expectedSysctls) {
 		t.Error("Sysctls was stripped from spec")
+	}
+
+	if spec.TaskTemplate.ContainerSpec.PidsLimit == 0 {
+		t.Error("PidsLimit was stripped from spec")
+	}
+	if spec.TaskTemplate.ContainerSpec.PidsLimit != 300 {
+		t.Error("PidsLimit did not preserve the value from spec")
 	}
 
 	if spec.TaskTemplate.ContainerSpec.Privileges.CredentialSpec.Config != "someconfig" {
@@ -70,6 +78,10 @@ func TestAdjustForAPIVersion(t *testing.T) {
 	adjustForAPIVersion("1.29", spec)
 	if spec.TaskTemplate.ContainerSpec.Sysctls != nil {
 		t.Error("Sysctls was not stripped from spec")
+	}
+
+	if spec.TaskTemplate.ContainerSpec.PidsLimit != 0 {
+		t.Error("PidsLimit was not stripped from spec")
 	}
 
 	if spec.TaskTemplate.ContainerSpec.Privileges.CredentialSpec.Config != "" {